+++ auth/src/main/java/com/firebase/ui/auth/compose/configuration/AuthUIConfiguration.kt
@@ -18,8 +18,8 @@ import android.content.Context
 import java.util.Locale
 import com.google.firebase.auth.ActionCodeSettings
 import androidx.compose.ui.graphics.vector.ImageVector
-import com.firebase.ui.auth.compose.configuration.stringprovider.AuthUIStringProvider
-import com.firebase.ui.auth.compose.configuration.stringprovider.DefaultAuthUIStringProvider
+import com.firebase.ui.auth.compose.configuration.string_provider.AuthUIStringProvider
+import com.firebase.ui.auth.compose.configuration.string_provider.DefaultAuthUIStringProvider
 import com.firebase.ui.auth.compose.configuration.theme.AuthUITheme
 
 fun actionCodeSettings(block: ActionCodeSettings.Builder.() -> Unit) =
