+++ auth/src/main/java/com/firebase/ui/auth/compose/configuration/string_provider/AuthUIStringProviderSample.kt
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.firebase.ui.auth.compose.configuration.stringprovider
+package com.firebase.ui.auth.compose.configuration.string_provider
 
 import android.content.Context
 import com.firebase.ui.auth.compose.configuration.AuthProvider
