+++ auth/src/main/java/com/firebase/ui/auth/compose/configuration/validators/PasswordValidator.kt
@@ -14,7 +14,7 @@
 
 package com.firebase.ui.auth.compose.configuration.validators
 
-import com.firebase.ui.auth.compose.configuration.stringprovider.AuthUIStringProvider
+import com.firebase.ui.auth.compose.configuration.string_provider.AuthUIStringProvider
 import com.firebase.ui.auth.compose.configuration.PasswordRule
 
 internal class PasswordValidator(
