+++ auth/src/main/java/com/firebase/ui/auth/compose/ui/components/ErrorRecoveryDialog.kt
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package com.firebase.ui.auth.compose
+package com.firebase.ui.auth.compose.ui.components
 
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.MaterialTheme
@@ -22,7 +22,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.window.DialogProperties
-import com.firebase.ui.auth.compose.configuration.stringprovider.AuthUIStringProvider
+import com.firebase.ui.auth.compose.AuthException
+import com.firebase.ui.auth.compose.configuration.string_provider.AuthUIStringProvider
 
 /**
  * A composable dialog for displaying authentication errors with recovery options.
